fixed bug
fixed the bug where you delete the generated descriptor file, and reran the second time, it generates again, but stops at this.

This was supposed to help with regenerating deleted descriptors but was causing more problems than making it easy.
So now, don't delete the descriptors, and if you do, you have to go update the two .txt and .snapshot in Cache folder...